これで構成確定する

+++ pkg/utils/string.go
@@ -5,15 +5,15 @@ import (
 	"strings"
 )
 
-func StrToIntArr(v string) ([]int, error) {
+func StrToIntArr(v string) ([]int32, error) {
 	parts := strings.Split(v, ",")
-	arr := make([]int, 0, len(parts))
+	arr := make([]int32, 0, len(parts))
 	for _, val := range parts {
 		v, err := strconv.Atoi(val)
 		if err != nil {
 			return nil, err
 		}
-		arr = append(arr, v)
+		arr = append(arr, int32(v))
 	}
 	return arr, nil
 }
